move func to new file game.h and add the instruction command for terminal
