+++ Task3Game/kernel/cmd.h
@@ -15,12 +15,5 @@ void drawImage();
 // video
 void playVideo(int x, int y);
 
-//welcome screen
-void drawWelcomeScreen(int screen_w, int screen_h);
-//draw end game screen
-void drawWinGameScreen(int screen_w, int screen_h);
-
-//game
-void playGame();
 //data logging
 void logCommand(char command);
