+++ Task3Game/kernel/game.h
@@ -1,0 +1,10 @@
+void drawWelcomeScreen(int screen_w, int screen_h);
+//draw end game screen  --- done
+void drawWinGameScreen(int screen_w, int screen_h);
+
+void drawMazeScreen(int screen_w, int screen_h, int steps);
+
+
+void playGame();
+
+void drawImage();
